requirements.txt
azure-core==1.26.3
azure-cosmos==4.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
easygui==0.98.3
idna==3.4
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
tk==0.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
